% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bogota_financial_time_series}
\alias{bogota_financial_time_series}
\title{Daily COVID-19 case data for Bogota and Medellin and corresponding Google mobility data}
\format{
\subsection{\code{bogota_financial_time_series}}{

A data frame with 1348 rows and 10 columns:
\describe{
\item{date}{Daily}
\item{city}{Bogota or Medellin}
\item{cases}{Daily counts of COVID-19 cases}
\item{deaths}{Daily counts of COVID-19 deaths}
...
}
}
}
\usage{
bogota_financial_time_series
}
\description{
Daily COVID-19 case data for Bogota and Medellin and corresponding Google mobility data
}
\keyword{datasets}
