% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_colombia_cases_deaths_mobility}
\alias{covid_colombia_cases_deaths_mobility}
\title{Daily COVID-19 case data for Bogota and Medellin and corresponding Google mobility data}
\format{
\subsection{\code{covid_colombia_cases_deaths_mobility}}{

A data frame with 1348 rows and 10 columns:
\describe{
\item{date}{Daily}
\item{city}{Bogota or Medellin}
\item{cases}{Daily counts of COVID-19 cases}
\item{deaths}{Daily counts of COVID-19 deaths}
...
}
}
}
\source{
\url{https://github.com/TRACE-LAC/pet-epi-notebooks/tree/main}

\url{https://www.google.com/covid19/mobility/}
}
\usage{
covid_colombia_cases_deaths_mobility
}
\description{
Daily COVID-19 case data for Bogota and Medellin and corresponding Google mobility data
}
\keyword{datasets}
