% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidp-package.R
\docType{package}
\name{epidp-package}
\alias{epidp}
\alias{epidp-package}
\title{epidp: Estimating Epidemiological Parameters Of Interest Using Covariate Information}
\description{
This package uses Stan to estimate epidemiological quantities of interest.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ben18785.github.io/epidp/}
}

}
\author{
\strong{Maintainer}: Ben Lambert \email{ben.c.lambert@gmail.com}

Authors:
\itemize{
  \item Kris Parag \email{k.parag@imperial.ac.uk}
}

}
\keyword{internal}
