% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_epifilter}
\alias{fit_epifilter}
\title{Bayesian linear regression with Stan}
\usage{
fit_epifilter(N, C, wmax, w, ...)
}
\arguments{
\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}

\item{x}{Numeric vector of input values.}

\item{y}{Numeric vector of output values.}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}
}
\description{
Bayesian linear regression with Stan
}
