% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_epifilter_covariates}
\alias{fit_epifilter_covariates}
\title{Estimate the time-varying reproduction number from incidence data using covariate
data}
\usage{
fit_epifilter_covariates(N, C, w, X, is_sampling = TRUE, ...)
}
\arguments{
\item{N}{number of data points}

\item{C}{case counts}

\item{w}{discretised serial interval}

\item{X}{matrix of covariates (which likely should include a first column of 1s)
of dimensions N x (N_covariates + 1)}

\item{is_sampling}{Boolean indicating if sampling is to be used (if TRUE) or
optimisation (if FALSE). Defaults to TRUE}

\item{...}{Arguments passed to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}} (if \code{is_sampling=TRUE}) or
\code{\link[rstan:stanmodel-method-optimizing]{rstan::optimizing()}} (if \code{is_sampling=FALSE})}
}
\value{
a stanfit object
}
\description{
Estimate the time-varying reproduction number from incidence data using covariate
data
}
