% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_epifilter}
\alias{fit_epifilter}
\title{Estimate the time-varying reproduction number from incidence data}
\usage{
fit_epifilter(N, C, w, is_sampling = TRUE, ...)
}
\arguments{
\item{N}{number of data points}

\item{C}{case counts}

\item{w}{discretised serial interval}

\item{is_sampling}{Boolean indicating if sampling is to be used (if TRUE) or
optimisation (if FALSE). Defaults to TRUE}

\item{...}{Arguments passed to \code{\link[rstan:stanmodel-method-sampling]{rstan::sampling()}} (if \code{is_sampling=TRUE}) or
\code{\link[rstan:stanmodel-method-optimizing]{rstan::optimizing()}} (if \code{is_sampling=FALSE})}
}
\value{
a stanfit object
}
\description{
Estimate the time-varying reproduction number from incidence data
}
