% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_models.R
\name{fit_epifilter_covariates}
\alias{fit_epifilter_covariates}
\title{Estimate the time-varying reproduction number from incidence data using covariate
data}
\usage{
fit_epifilter_covariates(N, C, w, X, is_sampling = TRUE, ...)
}
\arguments{
\item{N}{number of data points}

\item{C}{case counts}

\item{w}{discretised serial interval}

\item{X}{matrix of covariates (which likely should include a first column of 1s)
of dimensions N x (N_covariates + 1)}

\item{...}{}
}
\value{
a stanfit object
}
\description{
Estimate the time-varying reproduction number from incidence data using covariate
data
}
